Refactor to Potential function.
